requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
imageio-ffmpeg==0.5.1
multidict==6.0.5
numpy==1.26.4
psycopg2-binary==2.9.9
py-cord==2.5.0
typing_extensions==4.12.2
yarl==1.9.4
